made every service asynchronous

+++ Cyber.Application/Services/AddressService.cs
@@ -6,6 +6,7 @@ using Cyber.Core.Helper;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -22,28 +23,40 @@ public class AddressService
         _mapper = mapper;
     }
 
-    public void AddAddress(AddAddressDto address)
+    public async Task AddAddress(AddAddressDto address, int userId)
     {
         var addressToAdd = _mapper.Map<Address>(address);
-        _genericAddressService.Add(addressToAdd);
+        addressToAdd.UserId = userId;
+        await _genericAddressService.Add(addressToAdd);
     }
 
-    public List<AddressDto> GetAll()
+    public async Task<List<AddressDto>> GetAll()
     {
-        var allAddresses = _genericAddressService.GetAll();
+        var allAddresses = await _genericAddressService.GetAll();
         var allAddressesToDto = _mapper.Map<List<AddressDto>>(allAddresses);
         return allAddressesToDto;
     }
 
-    public AddressDto GetById(int id)
+    public async Task<AddressDto> GetById(int id)
     {
-        var address = _genericAddressService.GetById(id);
+        var address = await _genericAddressService.GetById(id);
         var addressToDto = _mapper.Map<AddressDto>(address);
         return addressToDto;
     }
 
-    public void Delete(int id)
+    public async Task<List<AddressDto>> GetUserAddresses(int userId)
     {
-        _genericAddressService.Delete(id);
+        var addresses = await _genericAddressService.Filter(a => a.UserId == userId);
+        if (addresses == null) throw new ArgumentException("No addresses found");
+        var addressToDto = _mapper.Map<List<AddressDto>>(addresses);
+        return addressToDto;
+    } 
+
+    public async Task Delete(int userId, int id)
+    {
+        var address = await _genericAddressService.GetById(id);
+        if (address == null) throw new ArgumentException("Address not found");
+        if (address.UserId != userId) throw new UnauthorizedAccessException("You are not authorized to delete this address");
+        await _genericAddressService.Delete(id);
     }
 }
